requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart==0.0.6
openai>=1.12.0
requests==2.31.0
python-dotenv<=1.0.0